rename files and packages
